reafactor and cleanup
